requirements.txt
# Core dependencies
fastapi>=0.68.0
uvicorn>=0.15.0
python-bitcoinlib>=0.11.0
web5>=0.1.0

# AI/ML
torch>=1.9.0
numpy>=1.21.0
scikit-learn>=0.24.2
transformers>=4.11.0

# Security
cryptography>=3.4.7
PyJWT>=2.1.0
python-jose>=3.3.0
passlib>=1.7.4

# Monitoring
opentelemetry-api>=1.5.0
opentelemetry-sdk>=1.5.0
prometheus-client>=0.11.0

# Testing
pytest>=6.2.5
pytest-asyncio>=0.15.1
pytest-cov>=2.12.1

# Development
black>=21.7b0
isort>=5.9.3
mypy>=0.910
pylint>=2.9.6
